+++ src/main/java/net/sf/mpxj/Task.java
@@ -5058,6 +5058,30 @@ public final class Task extends ProjectEntity implements Comparable<Task>, Proje
    }
 
    /**
+    * Retrieve a "canonical" version of the Activity ID.
+    * This method handles the case where the Activity ID for
+    * a WBS entry will be prefixed with the Project ID.
+    * This method replaces the Project ID with the text "PROJECT",
+    * which allows WBS entries to be matched by Activity ID 
+    * across projects.
+    * 
+    * @return canonical Activity ID value
+    */
+   public String getCanonicalActivityID()
+   {
+      String activityID = getActivityID();
+      if (getSummary() && activityID != null)
+      {
+         String projectID = getParentFile().getProjectProperties().getProjectID();
+         if (projectID != null && activityID.startsWith(projectID))
+         {
+            activityID = "PROJECT" + activityID.substring(projectID.length());
+         }
+      }
+      return activityID;
+   }
+
+   /**
     * Retrieve the activity ID.
     *
     * @return activity ID value