+++ src/main/java/net/sf/mpxj/primavera/PrimaveraPMFileReader.java
@@ -56,6 +56,7 @@ import net.sf.mpxj.ActivityCodeValue;
 import net.sf.mpxj.ActivityStatus;
 import net.sf.mpxj.AssignmentField;
 import net.sf.mpxj.Availability;
+import net.sf.mpxj.ChildTaskContainer;
 import net.sf.mpxj.ConstraintType;
 import net.sf.mpxj.CostAccount;
 import net.sf.mpxj.CostAccountContainer;
@@ -254,6 +255,7 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
       result.sort((o1, o2) -> Boolean.compare(o2.getProjectProperties().getExportFlag(), o1.getProjectProperties().getExportFlag()));
 
       linkCrossProjectRelations(result);
+      populateBaselines(result);
 
       m_externalRelations = null;
 
@@ -286,6 +288,19 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
                Relation relation = successor.addPredecessor(predecessor, externalRelation.getType(), externalRelation.getLag());
                relation.setUniqueID(externalRelation.getUniqueID());
             }
+         }
+      }
+   }
+
+   private void populateBaselines(List<ProjectFile> projects)
+   {      
+      Map<Integer, ProjectFile> map = projects.stream().collect(Collectors.toMap(p -> p.getProjectProperties().getUniqueID(), p -> p));
+      for (ProjectFile project : projects)
+      {
+         ProjectFile baseline = map.get(project.getProjectProperties().getBaselineProjectUniqueID());
+         if (baseline != null)
+         {
+            project.setBaseline(baseline, t -> t.getCanonicalActivityID());
          }
       }
    }
@@ -656,10 +671,11 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
 
       properties.setCreationDate(project.getCreateDate());
       properties.setFinishDate(project.getFinishDate());
+      properties.setGUID(DatatypeConverter.parseUUID(project.getGUID()));
       properties.setName(project.getName());
       properties.setStartDate(project.getPlannedStartDate());
       properties.setStatusDate(project.getDataDate());
-      properties.setProjectTitle(project.getId());
+      properties.setProjectID(project.getId());
       properties.setUniqueID(project.getObjectId());
       properties.setExportFlag(false);
    }
@@ -1032,9 +1048,10 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
       }
 
       //
-      // Create hierarchical structure
+      // Create rough hierarchical structure (note parent tasks not populated yet)
       //
       m_projectFile.getChildTasks().clear();
+
       for (WBSType row : wbs)
       {
          Task task = m_projectFile.getTaskByUniqueID(row.getObjectId());
@@ -1047,16 +1064,14 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
          {
             m_projectFile.getChildTasks().remove(task);
             parentTask.getChildTasks().add(task);
-
-            if (m_wbsIsFullPath)
-            {
-               task.setWBS(parentTask.getWBS() + PrimaveraReader.DEFAULT_WBS_SEPARATOR + task.getWBS());
-            }
-
-            task.setActivityID(task.getWBS());
          }
       }
 
+      //
+      // Set the WBS and the Activity ID
+      //
+      populateWBS(m_projectFile.getProjectProperties().getProjectID(), m_projectFile);
+      
       //
       // Read Task entries and create tasks
       //
@@ -1210,6 +1225,17 @@ public final class PrimaveraPMFileReader extends AbstractProjectStreamReader
       updateDates();
    }
 
+   private void populateWBS(String prefix, ChildTaskContainer container)
+   {
+      for (Task task : container.getChildTasks())
+      {
+         String wbs = prefix + PrimaveraReader.DEFAULT_WBS_SEPARATOR + task.getWBS(); 
+         task.setWBS(wbs);
+         task.setActivityID(wbs);
+         populateWBS(wbs, task);
+      }
+   }
+   
    private Duration addDurations(Number... values)
    {
       return getDuration(NumberHelper.sumAsDouble(values));